chore: remove V2 from new graphs
