+++ src/client/src/views/dashboard/widgets/lightning/liquidityGraph.tsx
@@ -1,4 +1,4 @@
-import { HorizontalBarChartV2 } from '../../../../components/chart/HorizontalBarChartV2';
+import { HorizontalBarChart } from '../../../../components/chart/HorizontalBarChart';
 import { LoadingCard } from '../../../../components/loading/LoadingCard';
 import { useGetLiquidReportQuery } from '../../../../graphql/queries/__generated__/getChannelReport.generated';
 import { chartColors } from '../../../../styles/Themes';
@@ -61,7 +61,7 @@ export const LiquidityGraph = () => {
 
   return (
     <S.wrapper>
-      <HorizontalBarChartV2
+      <HorizontalBarChart
         data={liquidity}
         colorRange={[chartColors.green]}
         dataKey="Value"
